Seperated Serial into SerialWin32 and SerilLinux && implemented PortListener::GetPorts() on linux
